requirements.txt
appdirs==1.4.4
astroid==2.5.1
attrs==20.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bidict==0.21.2
blinker==1.4
certifi==2020.6.20
cffi==1.13.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.8
importlib-metadata==3.8.0
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.2.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.5
packaging==20.9
pipenv==2020.11.15
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.19
pylint==2.7.2
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.1
python-http-client==3.3.2
python-socketio==5.1.0
requests==2.22.0
sendgrid==6.6.0
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.4.3
starkbank-ecdsa==1.1.0
timeago==1.0.10
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.7
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.2.1
zipp==3.4.1
